Many tweaks and a bit of cleanup
We are now using the fixed layout-bin-packer installed via npm, not from
vendor. I also commented out all the old samples to quiet the view
deprications from the template compiler.

I added a `buffer` property that will add extra items to the beggining
and end so that some items are already rendered when initially starting
a scroll. I also tried an optimization to only call `rerender` when
`index` or `count` property from the bin packer change.

I also changed `didInitAttrs` to `didRecieveAttrs` so that we can handle
the layout changing. This also required clearing the cellMap and cells
when the layout changes. An example of how this works is in the `simple`
demo now where you can change the size of the items.

Lastly I added a bit of style to the items in the simple example so you
can see where the bounds are more easily.