[WIP] - Use private api for cellMap
Just testing this out. Will probably need to use private api for 1.13
support since the new api can't be added to an already released version.
Will make an PR in master exposing the `decode-each-key` function as a
public api and then link to it in the code so it can be changed later.