Fix dummy app
This fixes an issue related to the update to `layout-bin-packer`causing
an `NaN` index when calculating the starting index. It invaldiates `cells` when
rendering. This is needed to cause the cells to update during scrolling.

Basically how this works now is `rAF` monitors the scroll position and
if the positions change between frames it schedules a `rerender`. The
`willRender` hook then invalidates the `cells` CP which ensures the
correct items are rendered. Then glimmer takes over and renders the
correct cells.

I also guarded `contentElement` not being defined on initial render.

Fix js hint