Delegate to layout for calculating style
This commit  moves responsibility of generating css for a cell to the layout. This allows for layouts / positioning that are based on percentages rather than exact pixels.

To demonstrate how this works a new `percentage-columns-layout` is added that takes `items` `columns` and `itemHeight` as paramters where `columns` is an array of numbers that totals less than `100`. The layout then uses the `ShelfFirst` bin packer to position the items based on the percentages and calculates CSS that uses those percentages.

All the built in layouts use style generators that were moved into the `utils/style-generators`.