Fix percentage-columns-layout
Looks like a made a silly change that the tests didn't catch when first doing the percentage-columns-layout. This fixes that so the layout works properly. I also improved the tests so that it will catch it. The tests now check the top / left as well as the width property for the rendered items.

While I was in the code I changed a design decision that was bothering me. Instead of having to pass an array with the same number of items as the content to the layout, you just need to pass the number of items.

Finally I made the assert for the size of the columns more strict, since using a set of columns that were under 100 could cause the items to not fit into rows as desired. I fixed the tests that make sure we assert properly.