Fix tests broken in browsers that have visible scrollbars
This fixes the scrollbar calculation in the scroll top tests to work (at least on win browsers). The tests would always pass in OSX due to lack of scroll bars. The working tests highlighted some faulty logic in one of the tests where we were adding the size of a scroll bar to the height, even though there wasn't enough content to make a horizontal scroll bar.